migrate: Screen render path + Harness trajectory + http abort
Screen.render drives the diff through an Allocating writer borrowing the
render buffer (capacity retained across frames, still allocation-free in
steady state); std.fmt.format becomes writer.print; the non-blocking tty
write uses a raw libc-write helper that keeps the short-write/WouldBlock
backpressure contract the removed File.write had. Harness trajectory
write threads io and reads the serialized bytes via writer.buffered().
EventOrchestrator shutdown yield uses clock.sleep; main fcntl casts the
NONBLOCK flag at its real u32 width; the Lua http abort shuts the stream
down through its io-aware Stream.shutdown. zig build -Dsim=false is now
green.

+++ src/main.zig
@@ -83,7 +83,7 @@ fn appendStatusLineFmt(
 fn setNonBlocking(fd: posix.fd_t) error{Unexpected}!void {
     const flags = std.c.fcntl(fd, std.c.F.GETFL, @as(usize, 0));
     if (flags < 0) return error.Unexpected;
-    const nonblock_bit: usize = @bitCast(std.c.O{ .NONBLOCK = true });
+    const nonblock_bit: u32 = @bitCast(std.c.O{ .NONBLOCK = true });
     const new_flags: usize = @as(usize, @intCast(flags)) | nonblock_bit;
     if (std.c.fcntl(fd, std.c.F.SETFL, new_flags) < 0) return error.Unexpected;
 }
